transcriber/requirements.txt
fastapi
uvicorn
soundfile
python-multipart
boto3
python-dotenv
kafka-python
requests